แก้แสดงค่าในหน้า sucject_elearn

+++ SmartLearningManager/app/src/main/java/com/example/a747/smartlearningmanager/Subject_elearn.java
@@ -20,6 +20,7 @@ import android.widget.ImageView;
 import android.widget.TableLayout;
 import android.widget.TableRow;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -39,6 +40,7 @@ public class Subject_elearn extends AppCompatActivity {
     String subject;
     String status = "n";
     String telno;
+    String watch_status;
     int c_absent = 0;
     TextView subjCode;
     TextView lecturer;
@@ -149,17 +151,19 @@ public class Subject_elearn extends AppCompatActivity {
                         });
                         cell.setTextSize(TypedValue.COMPLEX_UNIT_SP, 18);
                         cell.setPadding(20, 20, 0, 20);
+                        watch_status = "";
                         if (c.getString("check_status").equalsIgnoreCase("N") && c.getString("check_watch_e").equalsIgnoreCase("null")) {
                             cell.setBackgroundColor(Color.parseColor("#FFFF99"));
                         }
                         if (c.getString("check_watch_e").equalsIgnoreCase("Y")) {
-                            cell.setBackgroundColor(Color.parseColor("#66FF66"));
+                            cell.setBackgroundColor(Color.parseColor("#F5F5F5"));
+                            watch_status = "Watched.";
                         }
                         if (c.getString("check_status").equalsIgnoreCase("N")) {
                             c_absent++;
                         }
                         absent.setText(String.valueOf(c_absent));
-                        cell.setText(c.getString("e_date") + "  " + c.getString("e_time"));
+                        cell.setText(c.getString("e_date") + "  " + c.getString("e_time").substring(0,5)+" "+watch_status);
                         cell.setLayoutParams(new TableRow.LayoutParams(TableRow.LayoutParams.MATCH_PARENT, TableRow.LayoutParams.WRAP_CONTENT, 1f));
                         row.addView(cell);
                         row.setLayoutParams(new TableLayout.LayoutParams(TableLayout.LayoutParams.MATCH_PARENT, TableLayout.LayoutParams.MATCH_PARENT, 1f));
@@ -218,32 +222,40 @@ public class Subject_elearn extends AppCompatActivity {
                     try {
                         Log.i("Setup", "Set past video detail...");
                         JSONArray data = new JSONArray(strJSON);
-                        SQLiteDatabase mydatabase = openOrCreateDatabase("Elearning", MODE_PRIVATE, null);
-                        mydatabase.execSQL("DROP TABLE IF EXISTS Elearning");
-                        mydatabase.execSQL("CREATE TABLE IF NOT EXISTS Elearning(subject_code VARCHAR, subject_name VARCHAR, subject_room VARCHAR, e_date VARCHAR, e_time VARCHAR, e_link VARCHAR);");
-                        TableLayout tl_datelist = (TableLayout) findViewById(R.id.tl_datelist);
-                        for (int i = 0; i < data.length(); i++) {
-                            JSONObject c = data.getJSONObject(i);
-                            mydatabase.execSQL("INSERT INTO Elearning VALUES('" + c.getString("subject_code") + "','" + c.getString("subject_name") + "','" + c.getString("subject_room") + "','" + c.getString("e_date") + "','" + c.getString("e_time") + "','" + c.getString("e_link") + "');");
-                            TableRow row = new TableRow(Subject_elearn.this);
-                            TextView cell = new TextView(Subject_elearn.this);
-                            cell.setId(i);
-                            cell.setClickable(true);
-                            cell.setOnClickListener(new View.OnClickListener() {
-                                @Override
-                                public void onClick(View v) {gotoVideo(v);
-                                }
-                            });
-                            cell.setTextSize(TypedValue.COMPLEX_UNIT_SP, 18);
-                            cell.setPadding(20, 20, 0, 20);
-                            cell.setText(c.getString("e_date") + "  " + c.getString("e_time"));
-                            cell.setLayoutParams(new TableRow.LayoutParams(TableRow.LayoutParams.MATCH_PARENT, TableRow.LayoutParams.WRAP_CONTENT, 1f));
-                            row.addView(cell);
-                            row.setLayoutParams(new TableLayout.LayoutParams(TableLayout.LayoutParams.MATCH_PARENT, TableLayout.LayoutParams.MATCH_PARENT, 1f));
-                            tl_datelist.addView(row);
-                            absent.setText("0");
-                            TextView tv_title_semester = (TextView) findViewById(R.id.title_semester);
-                            tv_title_semester.setText(" Semester "+c.getString("semester")+"/"+c.getString("enroll_year"));
+                        if(data.length() > 0){
+                            SQLiteDatabase mydatabase = openOrCreateDatabase("Elearning", MODE_PRIVATE, null);
+                            mydatabase.execSQL("DROP TABLE IF EXISTS Elearning");
+                            mydatabase.execSQL("CREATE TABLE IF NOT EXISTS Elearning(subject_code VARCHAR, subject_name VARCHAR, subject_room VARCHAR, e_date VARCHAR, e_time VARCHAR, e_link VARCHAR);");
+                            TableLayout tl_datelist = (TableLayout) findViewById(R.id.tl_datelist);
+                            for (int i = 0; i < data.length(); i++) {
+                                JSONObject c = data.getJSONObject(i);
+                                mydatabase.execSQL("INSERT INTO Elearning VALUES('" + c.getString("subject_code") + "','" + c.getString("subject_name") + "','" + c.getString("subject_room") + "','" + c.getString("e_date") + "','" + c.getString("e_time") + "','" + c.getString("e_link") + "');");
+                                TableRow row = new TableRow(Subject_elearn.this);
+                                TextView cell = new TextView(Subject_elearn.this);
+                                cell.setId(i);
+                                cell.setClickable(true);
+                                cell.setOnClickListener(new View.OnClickListener() {
+                                    @Override
+                                    public void onClick(View v) {
+                                        gotoVideo(v);
+                                    }
+                                });
+                                cell.setTextSize(TypedValue.COMPLEX_UNIT_SP, 18);
+                                cell.setPadding(20, 20, 0, 20);
+                                cell.setText(c.getString("e_date") + "  " + c.getString("e_time"));
+                                cell.setLayoutParams(new TableRow.LayoutParams(TableRow.LayoutParams.MATCH_PARENT, TableRow.LayoutParams.WRAP_CONTENT, 1f));
+                                row.addView(cell);
+                                row.setLayoutParams(new TableLayout.LayoutParams(TableLayout.LayoutParams.MATCH_PARENT, TableLayout.LayoutParams.MATCH_PARENT, 1f));
+                                tl_datelist.addView(row);
+                                absent.setText("0");
+                                TextView tv_title_semester = (TextView) findViewById(R.id.title_semester);
+                                tv_title_semester.setText(" Semester " + c.getString("semester") + "/" + c.getString("enroll_year"));
+                            }
+                        }else{
+                            Toast.makeText(getApplicationContext(),"Not found", Toast.LENGTH_SHORT).show();
+                            status = "n";
+                            c_absent = 0;
+                            getSubjectVideo(subject);
                         }
                     } catch (Exception e) {
                         e.printStackTrace();
